move file_io under utilities, and restructure utilities
