+++ scraper/utilities/file_io/cazy_dictionary.json
